Web: Finish removing *_EXTENSIONS ifdefs from front end (minus grammar)
The grammar and scanner will be another step, as will the SPIR-V backend.
This makes heavy use of methods #ifdef'd to return false as a global way
of turning off code, relying on C++ DCE to do the rest.
There is a net reduction of over 50KB, or 50% of the target size.
There is a net reduction of 125 #ifdefs.